fate/gif: Remove nonsense requirement
It seems as if it was intended to declare fate-gif-color as prerequisite
of the fate-gifenc% tests. Yet the latter do not need anything from
the former, so this would be unnecessary. Furthermore, given that this
line has no associated recipe, it actually cancels implicit rules for
fate-gifenc% instead of adding a prerequisite.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>